Using Stackage.
The cabal.config file is downloaded from Stackage.org (hackage
alternative, with a curated set of Haskell packages that are
guaranteed to build) as of January 7, 2015, from here:

http://www.stackage.org/lts/cabal.config

This also forces us to update yesod, yesod-core, yesod-auth,
yesod-static, yesod-form and yesod-test to current.  Another bunch of
changes to deal with the new packages is forthcoming.

At this point I'm also switching to GHC 7.8.3 (because Stackage
necessiates it), installed from Herbert's PPA:

https://launchpad.net/~hvr/+archive/ubuntu/ghc

Some messing with Debian apt configuration is in order.  The sandbox,
cabal-install, and yesod-bin have been rebuilt with ghc-7.8.3.

Next step is trying LTS Haskell.