Remove defaultCsrfMiddleware
Having trouble with CSRF middleware, so removing it.  Basically, you
can't log in to the site even though the POST contains an XSRF-TOKEN
cookie.  What is the deal with that?

This seems to be a risk step, but Yesod scaffold also appears to be
moving in this direction.  See

https://github.com/yesodweb/yesod-scaffold/pull/125, and
https://github.com/yesodweb/yesod-scaffold/commit/8afc7d873a6ab62c5bedb3054f9a2d7bb093

Also see gh-13.