Add urlRenderOverride.
This bit of code is commented out for now -- because I do not know how
to correctly deal with generated files (files in static/tmp/ and maybe
those in static/combined).

Three problems so far:

 -- Development setup will load files from static host.  I don't want
    that.

 -- Keter unpacks new bundles to temp/app-$n/ directories, where $n
    changes each time keter receives a bundle; so I can't configure
    proxy server (nginx) with a constant static path.

 -- Devel and `production` (heh heh) has two different sets of
    generated files; so this will look broken all the time.

I have tried (1) cookie-less static site; (2) redirecting /static/tmp
to the yesod app and /static/ to static directory which has the other
assets; (3) routing all /static/ to the Yesod app.  None of these
setups are perfect because none of them solve the above problems.