Add "static-root" setting.
Two TODO items down: (1) this change loads static site from
configuration; (2) this (sort of) solves the problem of turning off
static file serving for development; (3) static/tmp (and static/autogen,
if we must) could be served with an nginx routing trick.

Not the best solution, but this (so far) works.