make `stack test` etc "work".
Trouble is rooted in the fact that applications root URL appears from
all sorts of (mysterious, until I understand this better) places, which
are sort of hard to keep in my head (hence the mystery).

Seems to me that `appRoot` can be set from /config/settings.yml or
config/test-settings.yml; but APPROOT environment variable is not read
by test scaffold.